refactor: upgrade dlightjs to the latest alpha version
