--- frameworks/keyed/dlightjs/src/data.js
@@ -1,18 +1,0 @@
-let idCounter = 1
-
-const adjectives = ["pretty", "large", "big", "small", "tall", "short", "long", "handsome", "plain", "quaint", "clean", "elegant", "easy", "angry", "crazy", "helpful", "mushy", "odd", "unsightly", "adorable", "important", "inexpensive", "cheap", "expensive", "fancy"]
-const colours = ["red", "yellow", "blue", "green", "pink", "brown", "purple", "brown", "white", "black", "orange"]
-const nouns = ["table", "chair", "house", "bbq", "desk", "car", "pony", "cookie", "sandwich", "burger", "pizza", "mouse", "keyboard"]
-
-function _random(max) { return Math.round(Math.random() * 1000) % max };
-
-export function buildData(count) {
-  const data = new Array(count)
-  for (let i = 0; i < count; i++) {
-    data[i] = {
-      id: idCounter++,
-      label: `${adjectives[_random(adjectives.length)]} ${colours[_random(colours.length)]} ${nouns[_random(nouns.length)]}`
-    }
-  }
-  return data
-}
