+++ frameworks/non-keyed/dlightjs/package.json
@@ -18,15 +18,15 @@
     "url": "https://github.com/dlight-js/dlight.git"
   },
   "dependencies": {
-    "@dlightjs/dlight": "1.0.0-alpha.1"
+    "@dlightjs/dlight": "^1.0.0-alpha.18"
   },
   "devDependencies": {
-    "babel-preset-dlight": "1.0.0-alpha.1",
     "@babel/core": "7.15.0",
     "@babel/plugin-proposal-class-properties": "^7.18.6",
-    "@babel/plugin-proposal-decorators": "^7.23.0",
+    "@babel/plugin-proposal-decorators": "^7.23.6",
     "@rollup/plugin-babel": "5.3.0",
     "@rollup/plugin-node-resolve": "13.0.4",
+    "babel-preset-dlight": "1.0.0-alpha.13",
     "rollup": "2.52.3",
     "rollup-plugin-terser": "7.0.2"
   }
